+++ spec/spec_helper.rb
@@ -1,6 +1,13 @@
 # frozen_string_literal: true
 
-require 'rspec'
+require 'simplecov'
+SimpleCov.start
+
+if ENV['CI'] == 'true'
+  require 'codecov'
+  SimpleCov.formatter = SimpleCov::Formatter::Codecov
+end
+
 require 'wrata_api'
 
 # @return [String] Default project for api tests